adaptation de la base de données aux id en bigserial

+++ tables/donnee_invalides.sql
@@ -2,18 +2,17 @@
 
 -- DROP TABLE public.donnee_invalides;
 
-CREATE TABLE public.donnee_invalides
-(
-    id serial PRIMARY KEY,
-    id_bouee integer NOT NULL,
-    temperature numeric(4,2) NOT NULL,
-    salinite numeric(4,2) NOT NULL,
-    debit numeric(5,2) NOT NULL,
-    date_temps time without time zone NOT NULL,
-    longitude numeric(10,8) NOT NULL,
-    latitude numeric(10,8) NOT NULL,
-    batterie integer NOT NULL,
-    CONSTRAINT donnee_bouees_id_bouee_fkey FOREIGN KEY (id_bouee)
+CREATE TABLE public.donnee_invalides (
+    id            bigserial     PRIMARY KEY
+    , id_bouee    integer       NOT NULL
+    , temperature numeric(4,2)  NOT NULL
+    , salinite    numeric(4,2)  NOT NULL
+    , debit       numeric(5,2)  NOT NULL
+    , date_temps  time without time zone NOT NULL
+    , longitude   numeric(10,8) NOT NULL
+    , latitude    numeric(10,8) NOT NULL
+    , batterie    integer       NOT NULL
+    , CONSTRAINT donnee_bouees_id_bouee_fkey FOREIGN KEY (id_bouee)
         REFERENCES public.bouees (id) MATCH SIMPLE
         ON UPDATE NO ACTION
         ON DELETE NO ACTION
